Implement CallbackSet.
It's ... a set of callbacks. It can be used to implement direct-reference
for callback registration, rather than the key model currently used on Broadcaster.

We can't use d3.set because it only accepts string keys.
ES6 set isn't available yet, and we'd still have to extend it to enable calling the callbacks.